poshan-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
openpyxl==3.1.2
gunicorn==22.0.0
PyJWT==2.8.0